Implement nominal and telescopeframe like skyoffsetframe
