+++ ctapipe/coordinates/horizon_frame.py
@@ -1,0 +1,4 @@
+from astropy.coordinates import AltAz
+
+
+HorizonFrame = AltAz
